+++ src/App/frontend/src/codegen/schemas/layout.schema.v1.ts
@@ -10,6 +10,9 @@ export class LayoutSchemaV1 extends SchemaFile {
 
   async getSchema(): Promise<JSONSchema7> {
     generateCommonSchema();
+    const externallyAvailableComponentKeys = this.sortedKeys.filter(
+      (key) => this.configMap[key].config.availability === 'configurable',
+    );
     const out: JSONSchema7 = {
       $ref: '#/definitions/ILayoutFile',
       definitions: {
@@ -21,10 +24,10 @@ export class LayoutSchemaV1 extends SchemaFile {
               // in each of the component schemas (not normally possible with this code generator)
               title: 'Type',
               description: 'The component type',
-              enum: this.sortedKeys.map((key) => this.componentList[key]),
+              enum: externallyAvailableComponentKeys.map((key) => this.componentList[key]),
             },
           },
-          allOf: this.sortedKeys.map((key) => ({
+          allOf: externallyAvailableComponentKeys.map((key) => ({
             if: { properties: { type: { const: this.componentList[key] } } },
             then: { $ref: `#/definitions/Comp${key}` },
           })),
@@ -32,7 +35,7 @@ export class LayoutSchemaV1 extends SchemaFile {
       },
     };
 
-    for (const key of this.sortedKeys) {
+    for (const key of externallyAvailableComponentKeys) {
       const config = this.configMap[key];
       out.definitions = out.definitions || {};
       out.definitions[`Comp${key}`] = config.toJsonSchema();